+++ LibraryComponents/Classes/Core/AndesList/AndesListCell.swift
@@ -58,6 +58,7 @@ import UIKit
     var paddingTopChevron: CGFloat? = 0
     var paddingBottomChevron: CGFloat? = 0
     var numberOfLines: Int = 0
+    var thumbnailType: AndesThumbnailType?
 
     required init?(coder: NSCoder) {
         super.init(coder: coder)
@@ -100,6 +101,7 @@ import UIKit
         self.subtitleLbl.adjustsFontSizeToFitWidth = false
         self.subtitleLbl.lineBreakMode = .byTruncatingTail
         setupSeparatorStyle(separatorStyle: separatorStyle)
+        self.updateAccessibilityProperties(thumbnailType: customCell.thumbnailType)
     }
 
     func setupSeparatorStyle(separatorStyle: AndesSeparatorStyle) {
@@ -109,6 +111,12 @@ import UIKit
         case .singleLine:
             separatorView.backgroundColor = UIColor.Andes.gray100
         }
+    }
 
+    func updateAccessibilityProperties(thumbnailType: AndesThumbnailType?) {
+        self.thumbnailImg.isAccessibilityElement = true
+        self.thumbnailImg.accessibilityTraits = accessibilityTraits
+        guard let type = thumbnailType else { return }
+        self.thumbnailImg.accessibilityLabel = type == .icon ? "icon accessibility".localized() : "image accessibility".localized()
     }
 }